# openapi.json
{"openapi":"3.1.0","info":{"title":"Tourisafe - Smart Tourist Safety System","version":"0.1.0"},"paths":{"/register":{"post":{"summary":"Register","operationId":"register_register_post","parameters":[{"name":"username","in":"query","required":true,"schema":{"type":"string","title":"Username"}},{"name":"email","in":"query","required":true,"schema":{"type":"string","title":"Email"}},{"name":"password","in":"query","required":true,"schema":{"type":"string","title":"Password"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/login":{"post":{"summary":"Login","operationId":"login_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/generate_tourist_id":{"post":{"summary":"Generate Tourist Id","operationId":"generate_tourist_id_generate_tourist_id_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/fences":{"post":{"summary":"Create Fence","operationId":"create_fence_fences_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string","title":"Name"}},{"name":"fence_type","in":"query","required":true,"schema":{"type":"string","title":"Fence Type"}},{"name":"geojson","in":"query","required":true,"schema":{"type":"string","title":"Geojson"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/fences":{"get":{"summary":"List Fences","operationId":"list_fences_api_fences_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/check_point":{"get":{"summary":"Check Point","operationId":"check_point_check_point_get","parameters":[{"name":"lat","in":"query","required":true,"schema":{"type":"number","title":"Lat"}},{"name":"lon","in":"query","required":true,"schema":{"type":"number","title":"Lon"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/telemetry":{"post":{"summary":"Ingest Telemetry","operationId":"ingest_telemetry_telemetry_post","parameters":[{"name":"user_id","in":"query","required":false,"schema":{"type":"integer","title":"User Id"}},{"name":"lat","in":"query","required":false,"schema":{"type":"number","title":"Lat"}},{"name":"lon","in":"query","required":false,"schema":{"type":"number","title":"Lon"}},{"name":"timestamp","in":"query","required":false,"schema":{"type":"string","format":"date-time","title":"Timestamp"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sos":{"post":{"summary":"Sos","operationId":"sos_sos_post","parameters":[{"name":"user_id","in":"query","required":false,"schema":{"type":"integer","title":"User Id"}},{"name":"lat","in":"query","required":false,"schema":{"type":"number","title":"Lat"}},{"name":"lon","in":"query","required":false,"schema":{"type":"number","title":"Lon"}},{"name":"description","in":"query","required":false,"schema":{"type":"string","default":"SOS","title":"Description"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/incidents":{"get":{"summary":"Get Incidents","operationId":"get_incidents_api_incidents_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Index","operationId":"index__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_login_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_login_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/login"}}}}}}